pyproject.toml
[tool.poetry]
name = "azpyfunc"
version = "0.1.0"
description = ""
authors = ["cecilphillip <cecilphillip@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
azure-functions = "^1.0.5"
azf_wsgi = "^0.3.0"
bottle = "^0.12.18"
flask = "^1.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
